refactor, merge three function in three files into one class in facebook.js
