+++ src/routes/webhook.js
@@ -2,7 +2,7 @@ import 'babel-polyfill';
 import Router from 'koa-router';
 import config from '../config';
 import receivedPostback from '../utils/receivedPostback';
-import { receivedMessage } from '../utils/receivedMessage';
+import receivedMessage from '../utils/receivedMessage';
 
 const VERIFY_TOKEN = config.VERIFY_TOKEN;
