+++ src/utils/receivedPostback.js
@@ -1,11 +1,12 @@
 import request from 'request';
 import config from '../config';
 import findThreeNewVideos from '../models/findThreeNewVideos';
-import { receivedMessage, sendGenericMessageByArr } from './receivedMessage';
+import receivedMessage from './receivedMessage';
 import saveGetStartedData from '../models/saveGetStartedData';
-import sendTextMessage from './sendTextMessage';
+import FacebookOP from './facebook';
 
 const PAGE_TOKEN = config.PAGE_TOKEN;
+const fb = new FacebookOP();
 
 const startedConv = (senderID, timeOfPostback) => {
   let name = '';
@@ -21,7 +22,7 @@ const startedConv = (senderID, timeOfPostback) => {
       console.log(`Error: ${response.body.error}`);
     } else {
       name = JSON.parse(body);
-      sendTextMessage(senderID, `Hello ${name.first_name} do you have a pen?`);
+      fb.sendTextMessage(senderID, `Hello ${name.first_name} do you have a pen?`);
       try {
         saveGetStartedData(senderID, name.first_name, timeOfPostback);
       }
@@ -45,7 +46,7 @@ const receivedPostback = (event) => {
     receivedMessage(event);
   } else if (payload === 'NEW') {
     findThreeNewVideos().then(returnArr => {
-      sendGenericMessageByArr(senderID, returnArr);
+      fb.sendGenericMessageByArr(senderID, returnArr);
     });
   }  else {
     startedConv(senderID, timeOfPostback);
